requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==3.0.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
comm==0.2.2
debugpy==1.8.11
decorator==5.1.1
distro==1.9.0
durationpy==0.9
exceptiongroup==1.2.2
executing==2.1.0
google-auth==2.36.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
jiter==0.8.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kubernetes==31.0.0
langchain-core==0.3.21
langchain-openai==0.2.10
langsmith==0.1.147
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
nvidia-ml-py==12.560.30
oauthlib==3.2.2
ollama==0.4.6
openai==1.55.3
orjson==3.10.12
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pip-autoremove==0.10.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tenacity==9.0.0
tiktoken==0.8.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
websocket-client==1.8.0
